requirements.txt
Flask==3.0.3
fpdf==1.7.2
protobuf==5.28.3
PyPDF2==3.0.1
python-dotenv==1.0.1
google-generativeai==0.8.3


